+++ ledger/api/ledger/admin.py
@@ -1,5 +1,4 @@
 # Standard Library
-import logging
 from typing import Any
 
 # Third Party
@@ -11,8 +10,13 @@ from django.utils.translation import gettext_lazy as _
 # Alliance Auth
 from allianceauth.authentication.models import CharacterOwnership, UserProfile
 from allianceauth.eveonline.models import EveCorporationInfo
+from allianceauth.services.hooks import get_extension_logger
+
+# Alliance Auth (External Libs)
+from app_utils.logging import LoggerAddTag
 
 # AA Ledger
+from ledger import __title__
 from ledger.api import schema
 from ledger.api.helpers import (
     get_all_corporations_from_alliance,
@@ -22,7 +26,7 @@ from ledger.api.helpers import (
 from ledger.models.characteraudit import CharacterAudit
 from ledger.models.corporationaudit import CorporationAudit
 
-logger = logging.getLogger(__name__)
+logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 
 class LedgerAdminApiEndpoints: