+++ ledger/managers/character_audit_manager.py
@@ -1,13 +1,17 @@
-# Standard Library
-import logging
-
 # Django
 from django.db import models
 
 # Alliance Auth
 from allianceauth.eveonline.models import EveCharacter
+from allianceauth.services.hooks import get_extension_logger
 
-logger = logging.getLogger(__name__)
+# Alliance Auth (External Libs)
+from app_utils.logging import LoggerAddTag
+
+# AA Ledger
+from ledger import __title__
+
+logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 
 class CharacterAuditQuerySet(models.QuerySet):