+++ ledger/managers/corporation_audit_manager.py
@@ -1,10 +1,16 @@
-# Standard Library
-import logging
-
 # Django
 from django.db import models
 
-logger = logging.getLogger(__name__)
+# Alliance Auth
+from allianceauth.services.hooks import get_extension_logger
+
+# Alliance Auth (External Libs)
+from app_utils.logging import LoggerAddTag
+
+# AA Ledger
+from ledger import __title__
+
+logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 
 class CorporationAuditQuerySet(models.QuerySet):